dropbox_bucket_strava/requirements.txt
flask
gunicorn
dropbox
google-cloud-secret-manager
google-cloud-firestore
google-cloud-storage
google-auth
git+https://github.com/pathexplorer/fit2gpx.git@first
git+https://github.com/pathexplorer/gcp_actions.git@v.0.1.3
gpxpy
numpy
pandas
lxml
fitdecode
python-dotenv
datetime